# Conflicts:
#	back_coaching/src/Controller/Admin/SeanceCrudController.php

+++ back_coaching/src/Controller/Api/SeanceInscriptionController.php
@@ -4,6 +4,7 @@ namespace App\Controller\Api;
 
 use App\Entity\Seance;
 use App\Entity\Sportif;
+use App\Entity\HistoriqueAnnulation;
 use App\Enum\TypeSeance;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -39,12 +40,64 @@ class SeanceInscriptionController extends AbstractController
 
         // Si le sportif est déjà inscrit, on le désinscrit
         if ($participation) {
-            $this->entityManager->remove($participation);
-            $this->entityManager->flush();
-            return $this->json(['message' => 'Désinscription réussie'], Response::HTTP_OK);
+            // Vérifier si l'annulation est possible (24h avant)
+            if ($seance->canBeCancelledBySportif()) {
+                // Récupérer les informations nécessaires
+                $sportif = $participation->getSportif();
+                
+                // Récupérer l'historique d'annulation existant ou en créer un nouveau
+                $historiqueRepo = $this->entityManager->getRepository(HistoriqueAnnulation::class);
+                $historique = $historiqueRepo->findOneBy([
+                    'sportif' => $sportif,
+                    'seance' => $seance
+                ]);
+                
+                if (!$historique) {
+                    $historique = new HistoriqueAnnulation();
+                    $historique->setSportif($sportif);
+                    $historique->setSeance($seance);
+                    $historique->setNbAnnulation(1);
+                } else {
+                    // Incrémenter le compteur d'annulations existant
+                    $historique->setNbAnnulation($historique->getNbAnnulation() + 1);
+                }
+                
+                $historique->setDateAnnulation(new \DateTime());
+                $this->entityManager->persist($historique);
+                
+                // Supprimer la participation pour désinscrire le sportif
+                $this->entityManager->remove($participation);
+                $this->entityManager->flush();
+                
+                return $this->json([
+                    'message' => 'Désinscription réussie, annulation comptabilisée',
+                    'nbAnnulations' => $historique->getNbAnnulation()
+                ], Response::HTTP_OK);
+            } else {
+                // Annulation trop tardive, compter comme absence
+                $participation->setPresence(false);
+                $this->entityManager->flush();
+                
+                return $this->json([
+                    'message' => 'Désinscription acceptée mais comptabilisée comme absence car moins de 24h avant la séance',
+                ], Response::HTTP_OK);
+            }
         } 
         // Sinon, on l'inscrit
         else {
+            // Vérifier si le sportif a déjà annulé cette séance 2 fois
+            $historiqueRepo = $this->entityManager->getRepository(HistoriqueAnnulation::class);
+            $historique = $historiqueRepo->findOneBy([
+                'sportif' => $user,
+                'seance' => $seance
+            ]);
+            
+            if ($historique && $historique->getNbAnnulation() >= 2) {
+                return $this->json([
+                    'message' => 'Vous avez déjà annulé cette séance 2 fois, vous ne pouvez plus vous y inscrire.'
+                ], Response::HTTP_BAD_REQUEST);
+            }
+            
             // Vérifier la capacité de la séance selon son type
             $nombreSportifs = count($seance->getParticipations());
             $limitesSportifs = [
@@ -59,6 +112,26 @@ class SeanceInscriptionController extends AbstractController
                     'message' => sprintf('Cette séance %s est déjà complète (maximum %d sportifs)', 
                         $seance->getTypeSeance()->value, $limite)
                 ], Response::HTTP_BAD_REQUEST);
+            }
+
+            // Vérifier si le sportif a déjà une séance à un horaire conflictuel
+            $seanceRepository = $this->entityManager->getRepository(Seance::class);
+            if ($seanceRepository instanceof \App\Repository\SeanceRepository) {
+                $hasConflict = $seanceRepository->hasSportifSessionConflict($user, $seance->getDateHeure(), $seance->getId());
+                if ($hasConflict) {
+                    return $this->json([
+                        'message' => 'Vous avez déjà une séance programmée à un horaire conflictuel. ' .
+                        'Il doit y avoir au moins 1h30 entre deux séances.'
+                    ], Response::HTTP_BAD_REQUEST);
+                }
+                
+                // Vérifier si le sportif a déjà 3 séances réservées à l'avance
+                $seancesFutures = $seanceRepository->countFutureSessionsForSportif($user);
+                if ($seancesFutures >= 3) {
+                    return $this->json([
+                        'message' => 'Vous ne pouvez pas réserver plus de 3 séances à l\'avance.'
+                    ], Response::HTTP_BAD_REQUEST);
+                }
             }
 
             try {
